Refactor API 기능 부모로 이동
